requirements.txt
cassandra-driver==3.15.1
redis==2.10.6
six==1.11.0